[Backend] : Entities
Association between lifecyclestep & plant updated
